+++ automation_tasks_rs/src/encrypt_decrypt_with_ssh_key_mod/encrypt_decrypt_mod.rs
@@ -291,7 +291,7 @@ pub(crate) fn encrypt_symmetric(
         &nonce,
         secret_string_to_encrypt.expose_secret().as_bytes(),
     ) else {
-        panic!("{RED}Error: Encryption failed. {RESET}");
+        anyhow::bail!("{RED}Error: Encryption failed. {RESET}");
     };
 
     let mut encrypted_bytes = nonce.to_vec();
@@ -321,7 +321,7 @@ pub(crate) fn decrypt_symmetric(
         nonce,
         cipher_text,
     ) else {
-        panic!("{RED}Error: Decryption failed. {RESET}");
+        anyhow::bail!("{RED}Error: Decryption failed. {RESET}");
     };
     let secret_decrypted_string = SecretString::from(String::from_utf8(secret_decrypted_bytes)?);