+++ automation_tasks_rs/src/generic_functions_mod.rs
@@ -14,11 +14,11 @@ pub use cl::{BLUE, GREEN, RED, RESET, YELLOW};
 /// Initialize tracing to file logs/automation_tasks_rs.log
 ///
 /// The folder logs/ is in .gitignore and will not be committed.
-pub fn tracing_init() {
+pub fn tracing_init() -> anyhow::Result<()>{
     // uncomment this line to enable tracing to file
     // let file_appender = tracing_appender::rolling::daily("logs", "automation_tasks_rs.log");
 
-    let offset = time::UtcOffset::current_local_offset().expect("should get local offset!");
+    let offset = time::UtcOffset::current_local_offset()?;
     let timer = tracing_subscriber::fmt::time::OffsetTime::new(
         offset,
         time::macros::format_description!("[hour]:[minute]:[second].[subsecond digits:6]"),
@@ -33,11 +33,11 @@ pub fn tracing_init() {
     // Unset the environment variable RUST_LOG
     // unset RUST_LOG
     let filter = tracing_subscriber::EnvFilter::from_default_env()
-        .add_directive("iced=error".parse().unwrap_or_else(|e| panic!("{e}")))
-        .add_directive("wgpu_core=error".parse().unwrap_or_else(|e| panic!("{e}")))
-        .add_directive("iced_wgpu=error".parse().unwrap_or_else(|e| panic!("{e}")))
-        .add_directive("iced_winit=error".parse().unwrap_or_else(|e| panic!("{e}")))
-        .add_directive("wgpu_hal=error".parse().unwrap_or_else(|e| panic!("{e}")))
+        .add_directive("iced=error".parse()?)
+        .add_directive("wgpu_core=error".parse()?)
+        .add_directive("iced_wgpu=error".parse()?)
+        .add_directive("iced_winit=error".parse()?)
+        .add_directive("wgpu_hal=error".parse()?)
 ;
 
     tracing_subscriber::fmt()
@@ -49,4 +49,5 @@ pub fn tracing_init() {
         //.with_writer(file_appender)
         .with_env_filter(filter)
         .init();
+    Ok(())
 }